refactor: Extract action menu components
